Phase1/requirements.txt
google-cloud-storage
google-cloud-pubsub
pandas
requests
chardet
